Remote object programming tests should be of type HwVoqSwitchWithMultipleDsfNodesTest.
Summary:
For remote object programming, we need to program objects against a
remote switch id. This requires the presence of that switch id in original
dsf nodes config (so we can appropriately calculate max cores). This did
not matter when we were just hard coding max cores on init to 20. Bu now that
we calculate this dynamically, it matters and tests fail.

Reviewed By: nivinl

Differential Revision: D43279591

fbshipit-source-id: 61e0cd2445b419b977cdb4f278440600906ab014