Directly publish the state_machine_thread_heartbeat_missed counter from heartbeat watchdog thread
Summary:
Today we publish the 'state_machine_thread_heartbeat_missed' counter from publishStats()->publishFbagentCounters() but this function runs on the same 'main' thread as refresh(). So if the main thread gets blocked by refresh, we won't publish stats to ODS also.

In this diff, I move the publish of state_machine_thread_heartbeat_missed counter to  the thread that performs the heartbeat watchdog. So when the watchdog thread detects a miss in heartbeat, we'll still publish the counter to fb303, and hence eventually to ODS via fbagent monitoring.

Reviewed By: rajank7

Differential Revision:
D43215619

Privacy Context Container: L1125642

fbshipit-source-id: 3cbc2e931f74ee0c3522242866a65c5c7e4a75af