provide a way to hook hw switch event observer or call back in agent initializer
Summary:
HwSwitch::CallBack is really HwSwitch event observer. HwSwitch dispatches all events to SwSwitch using this interface. Provide a mechanism to allow SwSwitch to register another call back handler or HwSwitch event observer. This is to be used in tests. In tests an agent ensemble will hook itself to SwSwitch and receive (or be notified of) all HwSwitch events and will relay or dispatch as soon as possible to SwSwitch.

This is necessary in case if a test wants to wait for HwSwitch event (such as link up/down or packet rx) and is orchestrated by AgentEnsemble (and so running the whole agent under the hood) but has no way to know if the event occurred.

AgentEnsemble can provide a way to register such event observers and notify them after (or before) dispatching such events to SwSwitch.

To facilitate this, SwSwitch is extended to accept HwSwitch::Callback in its init method. Accordingly updated interface of AgentInitializer (and Initializer) to pass HwSwitch::Callback to SwSwitch's init method.

Reviewed By: jasmeetbagga

Differential Revision:
D43180696

Privacy Context Container: L1125642

fbshipit-source-id: 7f2224f1653813328273faf213cac47d80bce846