Fix attribute value logging in getAttr calls
Summary:
As titled. Previously we tried to convert to log-and-call model, but the attribute values are not known until the call is made.

Therefore, move logging back to post API call. Will figure out a way to separate logging the call (prior to API call) and logging the values (post API call).

Reviewed By: jasmeetbagga

Differential Revision:
D43365751

Privacy Context Container: L1125642

fbshipit-source-id: e7a1f0ac681fe308d23d35becb69a03dc8aca975