Cmis function to correct the configured power state and retur Module status readiness
Summary:
Implements the CMIS function to check if the current power configuration state is not same as desired one then change it to that (by setting and resetting LP mode) otherwise return true when module is in ready state otherwise return false.
This function will be used from state machine
The part of this function matches with setPowerOverride(), I will refactor to remove that code.

Reviewed By: harshitgulati18

Differential Revision:
D43337681

Privacy Context Container: L1125642

fbshipit-source-id: 3d484c9c6efaafbc1b763bb798f4bbc5157be8b9