New option and API to setup a throughput test for simulator
Summary:
Add a new test to validate setup of one neighbor each on 2 ports. This test is being added for a
specific purpose, to provide an option to get the control plane up and running for simulator
throughput tests on vendor side.

Vendors as of now have HwTests from OSS running in their environment against simulator. This
test along with the new flag added, will provide a control plane (from the test binary) to program
the forwarding plane (setup adjacencies, bringup ports etc.) of simulator. Once the dataplane is
programmed, network namespaces and containers could be used to setup an external
environment to inject traffic to the simulator. This can be used for throughput test of ASIC
simulator in vendor environment used for regular HwTests.

Reviewed By: jasmeetbagga

Differential Revision: D43023316

fbshipit-source-id: a26e5684b8d1b07417260748e6248cc0f32542b6