Transceiver function to check module power and readiness
Summary:
Created transceiver level function to do the same thing:  check if the module's power control configuration is same as the desired one. If it is  same then return true or false based on whether module is in ready state or  not. If the power controll config value is not same as desired one then configure it correctly and return false

Also this function runs  in the i2c controller event base thread as it will be called from state machine

Reviewed By: harshitgulati18

Differential Revision:
D43337678

Privacy Context Container: L1125642

fbshipit-source-id: 211b587973935ae2d6b05774d08e451c903f8751