Start collecting watermark level on fabric_tx queues
Summary:
As titled. This is the watermark stat of interest
on fabric tx queues. In particular on kamet platforms it
reflects the RCI (reduced credit indicator) watermark
value. This indicates congestion on the fabric device, telling
the downstream fabric device to reduce credits being sent to
it.

TODO: Get this value to a fb303 counter.

Reviewed By: nivinl

Differential Revision:
D43614352

Privacy Context Container: L1125642

fbshipit-source-id: e816b6a93c5f5bc5dfa982b3541e08a245175833