Update the queue stats lambda TODO comment
Summary:
We have a legitemate case where queues on different
HWs does not have the same stats available as others. Concretely
the stats reported on fabric tx queues are much more sparse
than those on non fabric ports. Hence not finding a stat key
is not a deficiency (as reflected by TODO) but a facet of the
varied HW

Reviewed By: nivinl

Differential Revision:
D43668390

Privacy Context Container: L1125642

fbshipit-source-id: 4a3fe76c7dd671511d23334bf18bbad97aec09d2