Fix OSS Build Errors on Show Host
Summary: In building FBOSS OSS, compilation errors were thrown because the implementation of the show host command for the FBOSS2 CLI was using C++20 methods, such as contains() for std::map. To resolve these errors, I check to see if ndpEntryPort is in portInfoEntries and portStatusEntries with find() instead of contains().

Reviewed By: simuthus-fb

Differential Revision:
D43744125

Privacy Context Container: L1125585

fbshipit-source-id: 32d009705c7c004d2c85d9ae5a2b1d993908f6a7