Print SLPC error for FB FPGA in case of MDIO failure
Summary: Added Minipack/Fuji SLPC parity error counter. When MDIOP failure happens then most of the time we have found this is due to a faulty link between IOB FPGA and the DOM FPGA so it gets resolved by re-seating of the PIM or power cycle. In this diff, whenever MDIO error happens, we will print the SLPC error counter which will indicate if the IOB->DOM link has an issue

Reviewed By: harshitgulati18

Differential Revision: D43859660

fbshipit-source-id: c53295396f9cbe05a6937922de632b2b8329a57c