fix run_test.py to properly exclude bad tests
Summary:
Currently, some bad tests are not properly excluded based on bad test list regex, for example,
"HwMirrorTest/0.HwMirrorStat" is not excluded even given bad test regex "HwMirrorTest.*". This is because
--gtest_filter seems matching based on wildcard, see https://fburl.com/j4qg5f0w. As a result, it will also
attempt to match ".", which is not equal to "/".

This diff resolves this issue by replacing .* with * in --gtest_filter argument.

Reviewed By: simuthus-fb

Differential Revision: D43992521

fbshipit-source-id: 9ef41300605cdae21f873a6ea9e4f2192dd5d99f