Handle various corner scenarios with fabric reachability
Summary:
Summarizing changes.

**HW properties**

This is what we get from HW reachability protocol = {hwswitchId, hwportId}.
We use dsfNode  struct in cfg and access it using **hwswitchId** to fill in hwSwitchName.
We use **hwportId** to access the platform mapping and populate hwPortName.

**Expected Neighbor Properties**

This is what we get from HW reachability protocol = {switchName, portName}.
We use dsfNode struct in cfg  and access using **switchName** to fill in expectedSwitchId.
We use **portName** to access the platform mapping and populate expectedPortId.

**Corner scenario**
Now if the endpoint is unattached, we can't access platform mapping and hence can't find the expectedPortId.We can't find hwPortName either. Handle that.

** Other changes done**
Add dsfNode map which is keyed on the switchId as well. (to derive the hwSwitchName).

**Test Changes**
* New test case to handle unattached endpoint.
* Also added a test case where expected neighbor info doesn't match with HW info at all.

Reviewed By: jasmeetbagga

Differential Revision:
D43762651

Privacy Context Container: L1125642

fbshipit-source-id: d0f5f8fb0ef29956993e2de99499200d7ddc5f01