Add Support for Local Options
Summary: I added local option support for the FBOSS2 CLI. Developers of CLI commands can populate a vector of LocalOption structs, specifying the flag and the help message of each option, in the specific Traits struct of the command. The local options will then be bubbled via the tree of Command structs in CmdList so that CmdSubcommands can register the local options to the appropriate CLI::App object.

Reviewed By: peygar

Differential Revision: D43933847

fbshipit-source-id: aa57d20155d90fda6db017197853fb0724c41d8b