rackmon2: Use register level locking instead of a list level
Summary:
Goal is to allow for clients to retrieve monitoring data without having to wait excessively long for registers to get updated which can take seconds to complete.

This is accomplished by making the lock more fine-grained by moving it from protecting the register list to the register history itself.

Next in order to provide atomic updates of the registers without locking the register for an excessive period of time (10-100s of ms), we use a shadow buffer to read the registers and swap active/inactive buffers after read.

While we had minimal impact to the 90-p and below numbers (within error margin) our 99-p and max numbers are improved by a factor of 50%.

Imported from: D41477492

Reviewed By: tao-ren

Differential Revision: D43926941

fbshipit-source-id: a8b02b51c1aa3e488af1002328f469fe168ba464