Fix Object Arg Type Discrepancies in CLI Commands
Summary:
There are several commands in the FBOSS2 CLI in which the ObjectArgTypeId is not initialized to the correct enum in the command traits.

An example is show port, which is already resolved in D44122300.

For these commands, the ObjectArgTypeId in the command traits is set to the correct value. Most of these corrections are based on the enum that was passed directly to the command tree before D44122300, except for show bgp table fbnet, where the wrong enum was passed into the tree.

Moreover, for show bgp stream subscriber post-policy, the wrong class is passed into the command and arg type handlers. This typo is corrected.

Reviewed By: peygar

Differential Revision:
D44142376

Privacy Context Container: L1125585

fbshipit-source-id: bf88378595133bdebab7911c7e37d0c94f55205d