Warm boot downgrade support for TAJO
Summary:
Use the newly added switch extn attribute on TAJO to specify the SDK version(major/minor/dot -
version) to downgrade to, which will allow organizing warm boot data by the running SDK in a
format that can be parsed by the target downgrade SDK version. As of now, this is being added
as a fallback option in case any major issues are discovered during the rollout of SDK1.42.8. This
lets us downgrade using warmboot from SDK1.42.8 to SDk1.42.1.6 which is currently running in
the fleet.

Reviewed By: zechengh09

Differential Revision: D43780317

fbshipit-source-id: 0de0e95e55dd5a3d2be3c4f1bf6b36bf378b4bff