Implement Start Packet Capture (PCAP) Command
Summary: I implemented a new command, start pcap, for the FBOSS2 CLI. The command will make Thrift calls to FBOSS Agent to start capturing packets. The user can specify the name of the packet capture, the maximum number of packets to capture, and the direction of the packets to capture via local options.

Reviewed By: zechengh09

Differential Revision: D44152027

fbshipit-source-id: adbc402e6cac4081a2d7b3e5346fe02d211fc06a