run_test.py: accept regular expressions from filter
Summary:
Today, run_test.py accepts single filter and passes it on to gtest_filter.
However, gtest_filter follows this synatx:

[POSTIVE_PATTERNS(s)]:-[NEGATIVE_PATTERN(s)]

i.e. it is not possible to mix positive and negative patterns.
However, we typically have a set of positive/negative patterns per feature.

Thus, extend run_test.py to accept such filters from a file, and then pass it
to gtest_filter one at a time to compile the list of tests to run.

Reviewed By: nivinl

Differential Revision: D44271800

fbshipit-source-id: fa69c0744c8aef439284a7d97d251cf6514299fb