Pass in ECN bit values to be used for AQM tests instead of bool
Summary:
Instead of the bool isEcn, make the AQM test accept ECN field value, which will
be set to indicate if its nonECT, ECT0 or ECT1 traffic that needs to be tested. This
will be used in the next diff in the stack to add a new test to validate ECT0 and
ECT1 traffic.

Reviewed By: shri-khare

Differential Revision: D44355523

fbshipit-source-id: 78ce676fbf58347fd9e673a41c9452e331aa0e7b