Fix invariant tests ECMP setup with prod conifg
Summary:
Previously when using prod config, ECMP routes are not properly setup because ECMP  helper would add offsets to nextHops' IP. This is due to the limitation of test config, where subnet IP is the same for all nextHops. Offset is needed for multiple Ecmp nextHops.

Now with Rsw Prod config, the subnetIP is the uplink IP we need to resolve to. With offset added, incorrect nextHop IPs are resolved. Routes pointing to the correct IP is set to dropped.

To fix this, add the option of not adding offsets, which should be used with prod config.

Differential Revision:
D44185165

Privacy Context Container: L1125642

fbshipit-source-id: 98a66be2a239d794daab97ba9a46f1cccf366626