fix ecmp trunk test
Summary:
as titled.

HwTests don't have NDP mechanism, so when link is down, fast ECMP shrink will happen but neighbor should also go down which is indicated in setup post warm boot for a test where the neighbor are expected to come up again.

this bug in test was undetected   so far but oper delta based state processing revealed it.

Reviewed By: zechengh09

Differential Revision: D44360049

fbshipit-source-id: d509a7fd371f9a279759f9cfc0d2327595350508