Resolve Redundant Show L2 Command
Summary: The show l2 FBOSS2 command, which was added to the CLI in D43381156, is redundant because there is an existing command that serves the same purpose, show mac details. As a result, I modified the implementation of show l2 such that it presents a redirect message telling users to run show mac details instead to get the L2 table.

Reviewed By: zechengh09

Differential Revision: D44375899

fbshipit-source-id: bcf604479c3cbee272444b8f159adf2ce9cc1ca0