change the queue id based on the platform
Summary:
As titled, change the queue ID based on the platform.

Given that the queue ID has to be chosen based on the platform, this diff adds HwAsic as an argument to most of the OmypicQos Utils which triggers changes to other HW tests to get it compiled.

NOTE: This diff do not change any underlying behavior. The queue ID will remain the same for all existing platforms but begins leveraging the olympic utilities based on the queue type and by platform.

Reviewed By: shri-khare

Differential Revision: D44451803

fbshipit-source-id: 1c4e86a643cb59e0a5b6fd78a95f2853037dd874