When choosing safe port profile ids for J2c, stick to speed set in chip config
Summary:
- J2c does not yet support dynamic port speed change
- In the Hw test chip config, we have set the port speed to 100G for NIF port
- In the J2c profiles we support both 100G and 200G for NIF ports
Current logic for picking safe profiles picks the max available speed. This
means we pick 200G. The chip on init though comes up with 100G ports. So
during applyConfig, we end up changing the config which crashes.

This flow bites us when running Agent{Test, Benchmarks} as there we create
the config w/o accounting for the port speed ports came up with. In HwTests
OTOH we factor it in.
For a ongoing deployment, we will need to generate J2c chip config with 200G
downlink ports. At that time, I will just have hw test and benchmark configs
use that configuration and get rid of this code.

Reviewed By: nivinl

Differential Revision: D44529806

fbshipit-source-id: fac8ea417f2b5f17d8d519a9b13f66290ea0d7a6