Fix initialConfig() for prodInvariant tests
Summary:
As titled. InitialConfig is non-reentrant in prod invariant tests - we determine test/prod config by checking whether port is empty. If yes, treat it as test config and populate ports.

When initialConfig() is invoked again, ports are non-empty and hence treated as prod config. This is one of the reasons why the previous fix is breaking test config run.

Since we need a flag for setting up Ecmp routes, use the same flag to guard initialConfig().

Differential Revision:
D44695800

Privacy Context Container: L1125642

fbshipit-source-id: 19be0d374a36ed589ab7d674e4f51d16ab0566f9