Find application advertising field on a given host lane
Summary: Module can support a single application on multiple data paths. Today we configure a single application on the module so we didn't need to find the application field on any lane other than lane 0. With multi-port optics, we need to find if the module supports a given application on a given start lane.

Reviewed By: rajank7

Differential Revision:
D44393701

Privacy Context Container: L1125642

fbshipit-source-id: 19ebcaf373a8602aedeb61b05ffdb8cdc49d4862