Show switch drain state
Summary:
Overview
=======
Today we have "fboss amidrained" command which shows whether the switch is drained or not. Implementing the same in FBOSS2 CLI.

**Design**
========
Unlike in "fboss amidrained" where we query coop to get the switch
drain state, in FBOSS2 CLI we are going to query FBOSS local drainer which is the correct SoT to get switch drain state.

**Command**: "fboss2 show state"
**Output**: One of the strings in { DRAINED | SOFT_DRAINED | WARM_DRAINED |
UNDRAINED }

Implementation
=============
**This Diff:**
Skeleton code to introduce the command.

**Followup diffs:**
1. Create thrift client to query fboss_local_drainer
2. Query local drainer and print the correct output
3. UTs

Reviewed By: jasmeetbagga

Differential Revision: D44798462

fbshipit-source-id: 6dea4cc381d050b8210797ea58b6804a1625e725