Move `fboss2 show state drain` under facebook
Summary:
As titled to address comments in D44805517 and the rationale behind this move.

In short, this command queries fboss_local_drainer and as we do NOT open source fboss_local_drainer moving this command to facebook directory.

Also, we (zechengh09, shri-khare, jasmeetbagga) we decided use `fboss2 show state drain` to show drain state of the switch. So that we can enhance `fboss2 show state` parent command to show other state of the switch. For example `fboss2 show state switch` to show agent swswitch state.

Reviewed By: zechengh09

Differential Revision: D44859998

fbshipit-source-id: e47537bca94a3c3485ccd9a31a0276756a5a258d