Fix Parsing Bugs in Show Mirror
Summary:
I fixed a error in parsing IP addresses of mirror tunnels. Inside a tunnel object, the IP address is represented as another object (instead of a string) that needs to be further handled and converted into a string of readable format.

Also, because mirrors are being moved to mirrorMaps (instead of mirrorMap) as part of supporting multi-NPU designs, I updated the show mirror implementation to make the correct Thrift call and correctly parse the newly structured JSON to get the mirror information.

Reviewed By: zechengh09

Differential Revision: D44720476

fbshipit-source-id: 0b9e41a106558bb0e9bfce8e735837ecba99c8d5