Remove switchType() config setting in SaiPhyManager
Summary: As part of multi npu support, we are removing existing switchType and switchId fields in config and replacing it with switchIdToSwitchType map. SaiPhyManager is setting the switchType but it does not seem to be used anywhere. I am removing the setting. Let me know if this is used anywhere.

Reviewed By: jasmeetbagga

Differential Revision: D44515862

fbshipit-source-id: bdd1f06bd03fc75bac5fbbbf22d54eb25c140e8d