systemPortRange change by looking at system port ranges of all voq switches
Summary:
as titled. Previously we were looking at only the first switch ID,
which is not correct. Fix that.
However, this uncovers another deficiency, where we are maintaining a single
sys port range at switchSettings level. We need to either make this setting
part of switchIdToSwitchInfo map, or get deprecate it completely

Reviewed By: msomasundaran

Differential Revision: D44992679

fbshipit-source-id: 05e81397422be34d7ff6af8ea21e86547b75fbbf