Disable credit WD as part of TX disable for VoQ switches
Summary:
VoQ switches will purge traffic in queue if it stays in the  queue for more than a
configured time. With TX disabled while performing tests, this feature can cause
drops in the queue and hence test can fail, examples are accuracy tests for
watermark and ecn/wred thresholds. So, as part of the TX disable setting, make
sure that credit WD is disabled/enabled.

Reviewed By: jasmeetbagga

Differential Revision: D44528062

fbshipit-source-id: 9770584b1aaa0479e5e07c91f467c1e94c03ca56