Add support for Credit watchdog switch SAI attribute
Summary:
This attribute is used to enable/disable credit watchdog for VoQ systems. For
tests with transmit disabled, traffic stays in the VoQ and will end up getting
purged if the timer (default 500ms in DNX) expires. Looking at avoiding this
purge to make tests work irrespective of how long traffic is in the queue. No
change for prod behavior.

Reviewed By: jasmeetbagga

Differential Revision: D44528060

fbshipit-source-id: bee3fb2fa8b5fef8062f3f498bd75c2cd46a1f32