Deprecate the pfcwatchdog recovery action global field in switchstate
Summary: Switchsettings has a single field object pfcWatchdogRecoveryAction at switchstate root. Deprecating this field so that we can be ready for multi npu switch state migration. The single field is a cached entry created out of the settings at port level. For now, i have kept the old api but it return value from port fields. As a long term cleanup, the values can be read and applied from port level. I leave this to the pfc owners

Reviewed By: nivinl

Differential Revision: D45006408

fbshipit-source-id: b6490420945558ebab76c7add58efe13ab359cc0