Change the num packet to send for watermark accuracy back to 100
Summary:
Watermark accuracy tests have the assumption that the number of packets
being TXed is such that 80% of it is more than enough to use up the existing
credits for the queue. With the number reduced to 10, this fails for TH and
hence changing the number of packets TXed back to 100. This was changed
to accommodate Jericho2, but now that credit WD is implemented for Jericho2,
packets will not get purged while in the queue and hence using a bigger number
like 100 should be okay.

Reviewed By: peygar

Differential Revision: D45099799

fbshipit-source-id: 8e42b49e90755cd42effc0ad24e623d0f0e45727