expect difference in default map in multi-map and corresponding map
Summary:
this is because the default map in multi-map may contain obsoleted objects.

consider the following sequence
- prod to trunk

the multi map has been populated and are now being used by agent

- trunk to prod

the multi map got dumped in switch state (same map and default map in m-Map) and reloaded by prod and prod continues to use default map, but because we use ThriftCows::fromThrift, m-Map has also been populated as prod reloads stated dumped by trunk even though it is untouched

- prod to trunk

the prod dumps the state. the latest objects are in map (not in default map of m-Map), now trunk reloads the state and it is possible objects have been manipulated in map (but corresponding default map ha not been touched at all, because prod doesn't recognize them).

Now this is where trunk must expect multi-map contents are obsoleted.

the problem exists
- because prod agent's thrift definition already has multi map and so thrift cow's fromThrift retains it across prod2trunk2prod transitions
- the transitions prod2trunk2prod2trunk may trigger it if objects are updated.

Reviewed By: zechengh09, msomasundaran

Differential Revision: D45100923

fbshipit-source-id: 3cb62c1b0f237e6618edc2b09d7d90a7acb8da56