Populate interfaceIDs for port for warmbooting to new agent
Summary:
InterfaceIDs field for ports is recently introduced, and is populated during
config application.

When we warmboot from old (minus this field) to new (with this field), init
sequence involves:
 SwSwitch::init:
    -> SaiSwitch::init
       -> construct SwitchState from warmboot file... no interfaceID for port here
       -> apply that state
    -> notifyStateObservers
       -> NeighborCache::stateUpdated creates cache and attempts to send probe pkts.
          This would crash as checkReachability would look up interface from
          port (which is not populated): P698587855

 SwSwitch::applyConfig:
    -> populate interfaceID for port

This is not a problem for subsequent warmboots as interfaceID for port will be
written to warmboot file and thus available for SaiSwitch::init.

Reviewed By: jasmeetbagga

Differential Revision:
D45112083

Privacy Context Container: L1125642

fbshipit-source-id: 2a2b7014be1e719fb7d07eecdb2e8a816f74d322