Add API to get switchID associated with a port
Summary:
As titled. This will be useful in various contexts
when we need to figure out swithcIds for objects associated with a
port. For now making this work for only VOQ switches. However
as we goto global port ID assignment, it will work for all switch types

Reviewed By: msomasundaran

Differential Revision:
D45060663

Privacy Context Container: L1125642

fbshipit-source-id: 1314ff4dd6acd74cf66b7b13898dce302ecbeb6d