QSFP test to verify portNameToHostLane map
Summary:
as titled. Added this check as part of existing port profile tests so that it gets validated for all profiles.

Will add portNameToMediaLane check in another diff as we don't store media lanes in PlatformMapping so no good source of truth to compare against

Reviewed By: rajank7

Differential Revision:
D45163471

Privacy Context Container: L1125642

fbshipit-source-id: 2660c496a9ea13a718e29a9dd488f18b8b9d4337