Fix configuredMediaLanes
Summary:
Previously i was assuming that startMediaLane = startHostLane. But this isn't correct. We could have this mapping for instance for 400G optics in 4x100G mode

[HostLanes 0,1 => MediaLane 0]
[HostLanes 2,3 => MediaLane 1]
[HostLanes 4,5 => MediaLane 2]
[HostLanes 6,7 => MediaLane 3]

In the previous diff, we started caching the mediaLaneAssignment field in ApplicationAdvertisingField. In this diff, I use that to return the correct list of configuredMediaLanes

Reviewed By: rajank7

Differential Revision:
D45140472

Privacy Context Container: L1125642

fbshipit-source-id: 40b7d2228bc1ca830194c46810aae6faf12caf41