Display only relevant channels in show transceiver CLI
Summary: We now have a portNameToMediaLanes field which contains channel Ids for each port. Instead of printing all channels when the command is used for a specific port, this diff will filter out and print the relevant channels only. No difference today as each optic is associated with a single port, but will be useful to see relevant information for upcoming multi port OSFP optics.

Reviewed By: birdsoup

Differential Revision: D45293513

fbshipit-source-id: 07b45eb4855b502cd0366f2096502661fd95fb63