Simplifed numMedia/numHostLanes
Summary:
Simplified these two functions to return the number of lanes based on the moduleMediaInterface.

Moving forward
numMedia/numHostLanes -> Static value based on the inserted transceiver.
portNameToHost/MediaLanes -> Dynamic value based on what has been configured on the inserted transceiver.

There are also some changes to existing UTs which are working right now because we always returned a fixed numMedia/numHostLanes on sfp/qsfp modules. Now that it is based on eeprom, changes had to be made to the UTs.

Reviewed By: birdsoup

Differential Revision:
D45196729

Privacy Context Container: L1125642

fbshipit-source-id: 7dae0fa7a0651affb0e31f52a283323bd091d03a