Abstract out a template method for getting delta b/w first mnpu maps
Summary:
As we migrate to mnpu paradigm, one of the steps is to
get away from seeding default key (switchId=0) entries. So
instead of assuming and getting delta based on switchId=0 map
use the first map. This will unblock work for VOQ and fabric
switches where switchId is != 0. In parallel we are working
on a true mnpu delta computation abstraction as well. Which
will replace these APIs

Reviewed By: zechengh09, msomasundaran

Differential Revision:
D45327820

Privacy Context Container: L1125642

fbshipit-source-id: 9993120d8cb787bd8e2a42c28d4b17ab37015d18