Pass portName to test helpers
Summary:
QSFP tests currently verify all lanes in transceiverInfo. Now that we store a map of portName to Media/HostLanes in TransceiverInfo, we should only verify the info specific to the port.
This diff is the first step towards that. Passing portName down to the helpers and utils. Will use it in the next diff

Reviewed By: rajank7

Differential Revision:
D45296472

Privacy Context Container: L1125642

fbshipit-source-id: 88ec3803f2b57f341b3436512c057db32d7b2028