Don't queue remove_transceiver event from ~QsfpModule - fix deadlock on wedge100/galaxy
Summary:
We are seeing many deadlocks in qsfp service on wedge100 and galaxy devices when - 1) there are continuous link flaps on a port 2) there are occasional i2c errors on that port. When these two events happen at the same time, we see a deadlock.

The reason is that today the QsfpModule destructor calls removeTransceiver, which tries to queue a state change on the state machine thread. All of this is called with transceivers_.wlock because in the end we need to remove the QsfpModule from the transceivers_ map. So the sequence is
i. grab transceivers_.wlock
ii. ~QsfpModule
iii. Destructor queues a state change (for removing transceiver) on the state machine evb

If there are link flaps also going around, we'll see the following sequence as well
i. syncPorts
ii. Syncports queues a active/inactive event on the state machine evb
iii. Above requires a transceivers_.rlock

The deadlock is evident in the above two sequences. In the first sequence, a wlock is taken first and then a blocking function is scheduled on an evb. In the second sequence, a blocking function is schedule on the evb and that requires a rlock.

In this diff I fix that deadlock by making sure the events on evb are not scheduled with a wlock to transceivers_. That is, I process the remove transceiver event first with a rlock, and then acquire a wlock just to delete the module from the map.

Reviewed By: rajank7

Differential Revision:
D44121406

Privacy Context Container: L1125642

fbshipit-source-id: 736db4644f9846c71ad6163671d52de52614efc3