Query for Uplink/Downlink Interface in COPP Invariant Test
Summary:
The root cause of consistent COPP invariant test failures on prod configs is that the test implementation queries for the first interface from the prod config, which is not necessarily an uplink/downlink interface.

Instead, the COPP test will query for the first uplink/downlink interface (interface where `vlanID >= 2000`) and send test packets to the addresses of this interface.

This change should help stabilize the COPP invariant test for prod configs.

Reviewed By: zechengh09

Differential Revision:
D45448701

Privacy Context Container: L1125642

fbshipit-source-id: 280cc26102bf879bd1c8b9a9deb2ab60b927d8a1