PFC enabled priorities to return ones with non-zero headroom
Summary:
Currently all PG IDs configured are returned in Port getPfcPriorities().
However, there are PG configs now for lossy PGs as well and we need to
avoid those from getPfcPriorities(). We can assume that if headroom
is zero, its a lossless PG and return only those in getPfcPriorities().

Implication: The findEnabledPfcPriorities() is currently used to populate PFC
priorities, however, on warm boot, these configs are not changed and hence
will not trigger a change for native Broadcom RTSWs and hence should be
safe. Tried a warm boot upgrade on test RTSW - P717561049, looked clean.

If we want to be conservative and avoid changes to existing getPfcPriorities(),
we could introduce a new API and state data to get getPfcLosslessPriorities()
and use it for SAI alone, will make that change if needed based on  review.

   OK : 5
   FAILED : 1
   SKIPPED : 0
   TIMEOUT : 0
```

Differential Revision: D45534149

fbshipit-source-id: e7b0ba2398ee7fb628e00c9a48f0492a030bf8d5