Add a flag to allow a lossless PG with 0 headroom
Summary:
In the next diff, we are looking at changing the meaning of the PFC priorities,
which should translate to priorities where PFC should be enabled, which are
the lossless priorities. When we do that, any PG which doesn't have a
headroom configured will be considered lossy. This makes the test for 0 PG
headroom fail and need special handling to ensure for the test, PG with 0
headroom is still considered lossless and lets the test pass. This flag is
specifically for that.

Differential Revision: D45591926

fbshipit-source-id: c773de17a35260fccc3652dd28891e6d86cb2031