Fixed ECMP Next Hop Computation for Load Balancing Invariant Test
Summary:
The load balancing invariant test was failing for prod configs because the next hop addresses in the ECMP group did not match the prefixes of the uplink interfaces.

For example, an uplink interface may have the prefix `2401:db00:e12e:3625::59/127`. Under the current load balancing test implementation, we increment a nonzero offset to `2401:db00:e12e:3625::59` to get next hop `2401:db00:e12e:3625::5a` (or an address that is numerically higher), which would not match the first 127 bits of `2401:db00:e12e:3625::59`. As a result, the test does not detect any loopback traffic.

I computed next hop addresses by flipping the last bit. I decremented `lastOctet` if it was odd (change last bit from 1 to 0) and incremented `lastOctet` if it was even (change last bit from 0 to 1). This computation logic is only applied when we run ProdInvariantTests for prod configs.

Reviewed By: zechengh09

Differential Revision:
D45512946

Privacy Context Container: L1125642

fbshipit-source-id: c8f8ba3304c8f65fa43ae6b8724e44e5b2099a4b