Only pick cabled ports in findAvailablePorts
Summary:
Now that we have got rid of agent config for qsfp tests, we can't rely on platformMapping to return all available ports. This is because there are ports that get subsumed when some other port operates in a profile. So the current logic will pick incompatible ports.
Easiest is to assume that we always work with cabled ports in the test. This is not optimal for unit testing the XPHYs but making this change to unblock build failures.

Reviewed By: srikrishnagopu

Differential Revision:
D45838558

Privacy Context Container: L1125642

fbshipit-source-id: 56b384bd722c80aa24001a48cc4d2ace09e783dc