hw test to add both host route and neighbor
Summary:
this is a follow up from S336038

When same IP prefix is programmed as route and host entry, warm boot should not remove route entry. This was the case prior to SDK fix D45756138

However it is possible same IP prefix may get programmed both as route and host entry and warm boot should succeed. The test is added to confirm that

- add both route and host entry to same prefix
- warm boot
  - warm boot should find both LPM and host entry for prefix

Reviewed By: daiwei1983

Differential Revision: D45952065

fbshipit-source-id: d6377e58a4c9613367845229cf3fc6a06aa4b436