Fixed show lldp
Summary:
As mentioned in the T153484884, "fboss2 show lldp" was broken on some RSWs.
This happened because RSW received LLDP messages on ports connected to TW
hosts.
And agent thrift API, did NOT populate optional systemDescription field in
LinkNeighborThrift structure. Hence putting the check here for that optional
thrift field.

Reviewed By: peygar

Differential Revision: D45994605

fbshipit-source-id: 6dc2a159d7308dfaab2dfd96bcd15a37567c1bc6