Set XoffSize only if there is a change in the headroom config
Summary:
As of now, XoffSize is being set multiple times for the same value, this is not
necessary, check if the value has changed from the store and initiate a set
only if there is some delta.

This was causing CS00012294895, verified that with this fix, the multi PG
application works with diff D46058554, shared this with Broadcom also to
help fix the issue better.

Differential Revision: D45723362

fbshipit-source-id: 5e59b5830a6fd6289ce2c81a3b31caf2bc41b5ed