Parameterize processNeighborDelta, useful later for intfs
Summary:
VOQ/Fabric switches require that the packets are not tagged with VLAN.
We are gradually enhancing the wedge_agent to handle tagged as well as untagged
packets.

Diff stack D45034065 was the first phase in that enhancement and has landed. It
included changes for:
 - VOQ switches to use port based RIFs
 - Neighbor cache, and some other components continued to be VLAN based.
 - In order to make those work with VOQ switches, we used VVlanID 0
   (pseudoVlan) for VOQ switches to popoulate SwitchState/Neighbor cache etc.
   data structures.
 - Neighbor (ARP/NDP) tables and Neighbor respone tables for VOQ switches
   based off RIFs. For NPU switches, the Neighbor tables continued to be based off
   VLANs as before.
 - SaiSwitch was enhanced to process delta for neighbor tables in RIFs.

The second phase (this diff stack) will unify the implementation for NPUs and
VOQs. It will include:
 - VOQ switches use port based RIFs, NPU switches use VLAN based RIFs as
   before.
 - Neighbor cache will be indexed on InterfaceID, so no need for pseudo Vlan.
 - Neighbor response tables, neighbor tables will be based off interfaces not vlans
 - Both BcmSwitch and SaiSwitch to process delta for neighbor tables in RIFs.
 - Warmboot support to migrate from current model to the new model

General approach followeed by this diff stack
 - Introduce data structures / methods for migrating to the new model without
   directly invoking them in the code path i.e. don't alter current behavior.
 - Later diffs will invoke the methods to alter the behavior.

This approach allows splitting changes into smaller diffs and yet to breaking
trunk.

This diff: as titled.
Allows later stacked diff to use this util for interfaces

Reviewed By: jasmeetbagga

Differential Revision:
D46058716

Privacy Context Container: L1125642

fbshipit-source-id: 9bb910ea2f72141b80aac4a44a764b78ee5383d0