Pass switchState to getTable() helper
Summary:
The neighborTables will migrate from VLAN to Interface. However, MAC Table
continues to reside in VLAN.

Start passing switchState to getTable() helper, and thus:
 - for MAC table, lookup passed VLAN.
 - for interface table (future diffs), lookup interface from VLAN.

This is OK because the only callers of getTable() are from LookupClassUpdater
and LookupClassRouteUpdater which VOQ switch (which does not have VLANs)
exercises.

In future, update the callsites to use interfaces as well.

Reviewed By: jasmeetbagga

Differential Revision:
D46169033

Privacy Context Container: L1125642

fbshipit-source-id: b35514e3d33c9f4ada3044f911bb3e53fd708881