Introduce CONNECTING state when trying to connect FSDB
Summary:
Currently Fsdb stream client has state DISCONNECTED, CONNECTED and CANCELLED. For subscription/publication that are unreachable, client state will bounce back and forth between CONNECTED and DISCONNECTED. This is due to we transition from DISCONNECTED to CONNECTED whenever we **try** to connect.

It's hard to distinguish whether the connection is already established or still in the process of trying. Therefore, we'll have another state of CONNECTING, indicating that we're trying to connect to FSDB.

State transition should look like
  DISCONNECTED -> CONNECTING: Trying to connect to FSDB
  CONNECTING -> DISCONNECTED: Failed to connect to FSDB
  CONNECTING -> CONNECTED: Initial response received from FSDB
  CONNECTED -> DISCONNECTED: FSDB unreachable after connected.
  Any state can transition to CANCELLED.

Reviewed By: peygar

Differential Revision:
D45804514

Privacy Context Container: L1125642

fbshipit-source-id: 06fb9e890daa19eb7b5e051219ef0d7c54704320